requirements.txt
# Stream processing
faust-streaming>=0.10.15
aiokafka>=0.10.0,<0.12.0

# Ingestion
websockets>=12.0
certifi>=2024.0.0
httpx>=0.27.0
google-api-python-client>=2.140.0
youtube-transcript-api>=0.6.2

# ML / Embeddings
sentence-transformers>=3.0.1
scikit-learn>=1.5.2
numpy>=1.26.4
torch>=2.1.0

# Redis
redis[hiredis]>=5.0.8

# Dashboard
streamlit>=1.38.0
streamlit-autorefresh>=1.0.1
plotly>=5.24.1
pandas>=2.2.3

# Utilities
python-dotenv>=1.0.1
pydantic>=2.9.2
loguru>=0.7.2
aiohttp>=3.10.5

# Testing
pytest>=8.3.3
pytest-asyncio>=0.24.0
respx>=0.21.1
